requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
pip==23.0.1
python-dotenv==1.0.0
requests==2.28.1
setuptools==65.5.0
terminaltables==3.1.10
urllib3==1.26.12
